feat: create util.rs
