+++ circuit/src/storage_proof.rs
@@ -8,9 +8,12 @@ use plonky2::{
     plonk::circuit_builder::CircuitBuilder,
 };
 
-use crate::circuit::{slice_to_field_elements, CircuitFragment, D, F};
 use crate::gadgets::is_const_less_than;
 use crate::inputs::CircuitInputs;
+use crate::{
+    circuit::{CircuitFragment, D, F},
+    util::slice_to_field_elements,
+};
 
 pub const MAX_PROOF_LEN: usize = 64;
 pub const PROOF_NODE_MAX_SIZE_F: usize = 73;
